Share pool_factory fixture
